break out main split view controller
